refactor: use new wrappers for documentation and completions
