+++ src/main/java/io/github/handofgod94/common/wrappers/AttributeInfo.java
@@ -1,49 +1,62 @@
-package io.github.handofgod94.schema;
+package io.github.handofgod94.common.wrappers;
 
 import io.github.handofgod94.common.XmlUtil;
 import java.util.Iterator;
 import java.util.Optional;
 import java.util.stream.Stream;
 import org.apache.xerces.xs.XSAnnotation;
-import org.apache.xerces.xs.XSElementDeclaration;
+import org.apache.xerces.xs.XSAttributeDeclaration;
+import org.apache.xerces.xs.XSAttributeUse;
+import org.apache.xerces.xs.XSSimpleTypeDefinition;
 import org.dom4j.Document;
 import org.dom4j.Element;
 import org.eclipse.lsp4j.CompletionItem;
 import org.eclipse.lsp4j.CompletionItemKind;
+import org.eclipse.lsp4j.InsertTextFormat;
 import org.eclipse.lsp4j.MarkupContent;
 import org.eclipse.lsp4j.MarkupKind;
 
-public class ElementInfo {
-  private XSElementDeclaration elementDeclaration;
+/**
+ * Wrapper class around {@link XSAttributeUse}.
+ * It accepts the {@link XSAttributeUse} and provides additional
+ * method to retrieve information required for completions and hovers.
+ */
+public class AttributeInfo {
+  private XSAttributeDeclaration attributeDeclaration;
   private String name;
-  private String namespace;
+  private XSAttributeUse attributeUse;
+  private XSSimpleTypeDefinition type;
 
-  public ElementInfo(XSElementDeclaration elementDeclaration) {
-    this.elementDeclaration = elementDeclaration;
-    name = elementDeclaration.getName();
-    namespace = elementDeclaration.getNamespace();
+  public AttributeInfo(XSAttributeUse attributeUse) {
+    this.attributeUse = attributeUse;
+    attributeDeclaration = attributeUse.getAttrDeclaration();
+    name = attributeDeclaration.getName();
+    type = attributeDeclaration.getTypeDefinition();
   }
 
   public CompletionItem toCompletionItem() {
-    CompletionItem item = new CompletionItem();
-    item.setLabel(name);
-    item.setDetail(namespace);
-    item.setDocumentation(toMarkupContent());
-    item.setKind(CompletionItemKind.Field);
-    return item;
+    CompletionItem info = new CompletionItem();
+    info.setLabel(name);
+    info.setDetail(type.getName());
+    info.setDocumentation(toMarkupContent());
+    info.setInsertTextFormat(InsertTextFormat.Snippet);
+    info.setInsertText(name + "=\"$1\"");
+    info.setKind(CompletionItemKind.Property);
+    return info;
   }
 
   public MarkupContent toMarkupContent() {
+
     MarkupContent content = new MarkupContent();
     content.setKind(MarkupKind.MARKDOWN);
 
     StringBuffer docBuffer = new StringBuffer();
-    String elementStr = String.format("**ELEMENT**: %s  \n", name);
-    docBuffer.append(elementStr);
+    String attrStr = String.format("**ATTRIBUTE**: %s  \n", name);
+    docBuffer.append(attrStr);
 
     String annotation =
       Stream
-        .of(Optional.ofNullable(elementDeclaration.getAnnotation()))
+        .of(Optional.ofNullable(attributeDeclaration.getAnnotation()))
         .filter(Optional::isPresent)
         .map(Optional::get)
         .map(XSAnnotation::getAnnotationString)
@@ -62,7 +75,7 @@ public class ElementInfo {
         if (documentation.attributeCount() > 0) {
           descKey =
             (documentation.attributeValue("source") != null) ?
-              documentation.attributeValue("source").toUpperCase() : descKey;
+                documentation.attributeValue("source").toUpperCase() : descKey;
         }
         String descValue = documentation.getTextTrim();
         String descriptionDoc = String.format("**%s**: %s  \n", descKey, descValue);
@@ -70,21 +83,28 @@ public class ElementInfo {
       }
     });
 
+    String typeDoc = String.format("**TYPE**: %s  \n", type.getName());
+    docBuffer.append(typeDoc);
     content.setValue(docBuffer.toString());
+
     return content;
   }
 
   // Getters
 
-  public XSElementDeclaration getElementDeclaration() {
-    return elementDeclaration;
+  public XSAttributeDeclaration getAttributeDeclaration() {
+    return attributeDeclaration;
   }
 
   public String getName() {
     return name;
   }
 
-  public String getNamespace() {
-    return namespace;
+  public XSAttributeUse getAttributeUse() {
+    return attributeUse;
+  }
+
+  public XSSimpleTypeDefinition getType() {
+    return type;
   }
 }
